pyproject.toml
[tool.poetry]
name = "bto_adpter"
version = "0.1.0"
description = ""
authors = ["duresafeyisa2022@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
rdflib = "^6.3.2"
typer = "^0.9.0"
pyyaml = "^6.0"
neo4j = ">=4.4,<5.0"  # Ensure compatibility with biocypher
biocypher = "^0.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
